% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{get_channel_handle}
\alias{get_channel_handle}
\title{Extracts channel handle from given URL}
\usage{
get_channel_handle(url)
}
\arguments{
\item{url}{A character string of the YouTube channel URL.}
}
\value{
A character string of the channel handle.
}
\description{
Extracts channel handle from given URL
}
\examples{
get_channel_handle("https://www.youtube.com/@oggyinformatics/")
}
