% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{get_channel_name}
\alias{get_channel_name}
\title{Extracts channel name from given URL}
\usage{
get_channel_name(url)
}
\arguments{
\item{url}{A character string of the YouTube channel URL.}
}
\value{
A character string of the channel name.
}
\description{
Extracts channel name from given URL
}
\examples{
get_channel_name("https://www.youtube.com/@oggyinformatics/")
}
