% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{get_subscriber_count}
\alias{get_subscriber_count}
\title{Extracts subscriber count from given URL}
\usage{
get_subscriber_count(url)
}
\arguments{
\item{url}{A character string of the YouTube channel URL.}
}
\value{
A character string of the subscriber count.
}
\description{
Extracts subscriber count from given URL
}
\examples{
get_subscriber_count("https://www.youtube.com/@oggyinformatics/")
}
