% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{get_text}
\alias{get_text}
\title{Extracts text from given URL}
\usage{
get_text(url)
}
\arguments{
\item{url}{A character string of the URL.}
}
\value{
A character string of the extracted text.
}
\description{
Extracts text from given URL
}
\examples{
get_text("https://www.youtube.com/@oggyinformatics/")
}
