% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{get_video_count}
\alias{get_video_count}
\title{Extracts video count from given URL}
\usage{
get_video_count(url)
}
\arguments{
\item{url}{A character string of the YouTube channel URL.}
}
\value{
A character string of the video count.
}
\description{
Extracts video count from given URL
}
\examples{
get_video_count("https://www.youtube.com/@oggyinformatics/")
}
